requirements.txt
tensorcircuit-ng
tensornetwork-ng
numpy
scipy
jax
jaxlib
quimb
optax
cotengra
